app2/requirements.txt
numpy==1.26.4
Flask==3.0.3
pillow==10.3.0
scipy
matplotlib
gunicorn
azure-storage-blob 
azure-identity